Index.js is now the correct file
